requirements.txt
numpy
Pillow
opencv-python
matplotlib
